+++ oggm/tests/__init__.py
@@ -9,6 +9,18 @@ from packaging.version import Version
 from oggm import cfg
 from oggm.utils import SAMPLE_DATA_COMMIT
 
+# The temperature bias prior file to use in the tests: this is the file the
+# OGGM v1.6 preprocessed directories were calibrated with (W5E5, RGI6). There
+# is no default in OGGM, it always has to be given explicitly.
+TEMP_BIAS_FILE_W5E5_RGI6 = ('https://cluster.klima.uni-bremen.de/~oggm/'
+                            'ref_mb_params/oggm_v1.6/'
+                            'w5e5_rgi6_perglacier_temp_bias_v2025.6.2.csv')
+
+# Regional averages of the geodetic observations. OGGM calibrates on the
+# per-glacier values only, but this is a useful reference in the tests.
+GEODETIC_MB_REGIONAL_AVG = ('https://cluster.klima.uni-bremen.de/~oggm/'
+                            'geodetic_ref_mb/hugonnet_2021_regional_avg.csv')
+
 # Some logic to see which environment we are running on
 
 # Matplotlib version changes plots, too