+++ oggm/tests/test_workflow.py
@@ -20,7 +20,7 @@ import oggm
 import oggm.cfg as cfg
 from oggm import workflow
 from oggm.utils import get_demo_file, write_centerlines_to_shape, ModelSettings
-from oggm.tests import mpl_image_compare
+from oggm.tests import mpl_image_compare, TEMP_BIAS_FILE_W5E5_RGI6
 from oggm.tests.funcs import get_test_dir, use_multiprocessing, characs_apply_func
 from oggm.shop import cru
 from oggm.core import flowline, gis, inversion, centerlines, massbalance
@@ -31,7 +31,7 @@ from oggm.core.massbalance import (MonthlyTIModel, MultipleFlowlineMassBalance,
 from oggm.core.flowline import (run_from_climate_data, run_random_climate,
                                 init_present_time_glacier, run_constant_climate, MixedBedFlowline)
 from oggm.core.inversion import get_inversion_volume
-from oggm.exceptions import InvalidWorkflowError
+from oggm.exceptions import InvalidWorkflowError, InvalidParamsError
 
 # Globals
 pytestmark = pytest.mark.test_env("workflow")
@@ -702,6 +702,7 @@ class TestGdirSettings:
                                      gdirs,
                                      overwrite_gdir=True,
                                      informed_threestep=True,
+                                     temp_bias_file_path=TEMP_BIAS_FILE_W5E5_RGI6,
                                      settings_filesuffix='_informed_threestep',
                                      )
         mb_calib_threestep = gdirs[0].read_yml('settings',
@@ -716,6 +717,15 @@ class TestGdirSettings:
                         mb_calib_threestep['temp_bias'],
                         atol=6e-2)
 
+        # the temp bias prior file is not optional
+        with pytest.raises(InvalidParamsError):
+            tasks.mb_calibration_from_geodetic_mb(
+                gdir,
+                overwrite_gdir=True,
+                informed_threestep=True,
+                settings_filesuffix='_informed_threestep',
+            )
+
         # recalibration without overwrite_gdir should raise an error
         prcp_fac_original = settings_informed_threestep['prcp_fac']
         with pytest.raises(InvalidWorkflowError):
@@ -724,6 +734,7 @@ class TestGdirSettings:
                                          gdirs,
                                          overwrite_gdir=False,
                                          informed_threestep=True,
+                                         temp_bias_file_path=TEMP_BIAS_FILE_W5E5_RGI6,
                                          settings_filesuffix='_informed_threestep',
                                          )
 
@@ -1010,6 +1021,7 @@ class TestGdirObservations:
                                      gdirs,
                                      overwrite_gdir=True,
                                      informed_threestep=True,
+                                     temp_bias_file_path=TEMP_BIAS_FILE_W5E5_RGI6,
                                      settings_filesuffix='_informed_threestep',
                                      )
         # check that observation was added to the observation files and that the
@@ -1039,6 +1051,7 @@ class TestGdirObservations:
                                      gdirs,
                                      overwrite_gdir=True,
                                      informed_threestep=True,
+                                     temp_bias_file_path=TEMP_BIAS_FILE_W5E5_RGI6,
                                      settings_filesuffix='_informed_threestep',
                                      observations_filesuffix='_hugonnet_adapted',
                                      use_observations_file=True,